layouts: correctly check for template existence
https://github.com/jpanther/congo/commit/4e4b470915171a5855f3f2ea25851995aace8b55
fixed the theme for building with Hugo v0.146.0 or later but it
doesn't correctly check for template existence. That is, if you set
`homepage.layout` to `profile` this theme will still load the `page`
layout.

Hugo's [templates.Exists](https://gohugo.io/functions/templates/exists/)
function tests for the existence of templates relative to the `layouts`
directory. The example code they provide uses `printf` to join the
template name with the `_partials` directory, which would be `partials`
instead for this theme.

```
{{ $partialPath := printf "headers/%s.html" .Type }}
{{ if templates.Exists ( printf "_partials/%s" $partialPath ) }}
  {{ partial $partialPath . }}
{{ else }}
  {{ partial "headers/default.html" . }}
{{ end }}
```